+++ src/DotRecast.Detour.TileCache/DtLayerMonotoneRegion.cs
@@ -7,8 +7,8 @@ namespace DotRecast.Detour.TileCache
         public const int DT_LAYER_MAX_NEIS = 16;
 
         public int area;
-        public List<int> neis = new List<int>(DT_LAYER_MAX_NEIS);
-        public int regId;
-        public int areaId;
+        public List<byte> neis = new List<byte>(DT_LAYER_MAX_NEIS);
+        public byte regId;
+        public byte areaId;
     };
 }