+++ src/DotRecast.Recast/RcLayerSweepSpan.cs
@@ -3,7 +3,7 @@
     public class RcLayerSweepSpan
     {
         public int ns; // number samples
-        public int id; // region id
-        public int nei; // neighbour id
+        public byte id; // region id
+        public byte nei; // neighbour id
     };
 }